refactor(core): relocate Tasks cluster into app/tasks.rs
