+++ src/app/key_handlers.rs
@@ -699,98 +699,6 @@ impl App {
         }
     }
 
-    /// Handle keys while the tasks panel is focused: `j`/`k` select (and preview
-    /// the selected task in the central pane), `PageUp`/`PageDown` scroll that
-    /// preview, `n` create, `e`/`Enter` open the central-pane editor, `Space`
-    /// cycles status, `r` opens the trigger-time action picker, `d` deletes,
-    /// `Esc` leaves. Searching is handled by the global `Ctrl+/`.
-    pub(crate) fn handle_task_list_key(&mut self, code: KeyCode) {
-        // Creating works regardless of whether the panel has entries.
-        if matches!(code, KeyCode::Char('n')) {
-            self.new_task_in_pane();
-            return;
-        }
-
-        let count = self.task_ui.filtered_task_indices.len();
-        match code {
-            KeyCode::Char('j') | KeyCode::Down
-                if count > 0 && self.task_ui.task_panel_index + 1 < count =>
-            {
-                self.task_ui.task_panel_index += 1;
-                self.refresh_task_view();
-            }
-            KeyCode::Char('j') | KeyCode::Down => {}
-            KeyCode::Char('k') | KeyCode::Up => {
-                self.task_ui.task_panel_index = self.task_ui.task_panel_index.saturating_sub(1);
-                self.refresh_task_view();
-            }
-            KeyCode::Esc => {
-                self.focus = InputFocus::SessionList;
-                self.task_ui.task_editor = None;
-            }
-            // Open the central-pane editor for the selected task. On an empty
-            // panel, start a new task instead.
-            KeyCode::Char('e') | KeyCode::Enter => self.enter_task_editor_or_new(count),
-            // Scroll the full-screen preview (the list itself uses j/k).
-            KeyCode::PageDown => self.scroll_task_preview(5),
-            KeyCode::PageUp => self.scroll_task_preview(-5),
-            KeyCode::Char(' ') => self.cycle_selected_task_status(),
-            // Open the trigger-time action picker (Send → session / Spawn new).
-            KeyCode::Char('r') => self.open_selected_task_action_picker(),
-            // Jump to the task's related session terminal (the spawned
-            // `<title> · #<id>` window, or a persisted Send target).
-            KeyCode::Char('o') => self.open_task_related_session(),
-            KeyCode::Char('d') => self.delete_selected_task(),
-            _ => {}
-        }
-    }
-
-    /// `e`/`Enter` on the tasks panel: open the central-pane editor for the
-    /// selection, or start a new task when the panel is empty.
-    fn enter_task_editor_or_new(&mut self, count: usize) {
-        if count == 0 {
-            self.new_task_in_pane();
-        } else {
-            self.enter_task_editor();
-        }
-    }
-
-    /// `r` on the tasks panel: open the trigger-time action picker for the
-    /// selected task (Send → session / Spawn new).
-    fn open_selected_task_action_picker(&mut self) {
-        if let Some(task) = self.selected_task().cloned() {
-            self.open_task_action_picker(&task);
-        }
-    }
-
-    /// Cycle the selected task's status (`Todo → InProgress → Done`) and refresh.
-    fn cycle_selected_task_status(&mut self) {
-        let Some(task) = self.selected_task().cloned() else {
-            return;
-        };
-        if let Err(e) = self.db.set_task_status(task.id, task.status.cycle()) {
-            error!("Failed to cycle task {} status: {e}", task.id);
-        }
-        self.refresh_tasks();
-        self.sync_task_editor();
-    }
-
-    /// Soft-delete the selected task, then clamp the selection and refresh.
-    fn delete_selected_task(&mut self) {
-        let Some(id) = self.selected_task().map(|t| t.id) else {
-            return;
-        };
-        if let Err(e) = self.db.soft_delete_task(id) {
-            error!("Failed to delete task {id}: {e}");
-        }
-        self.refresh_tasks();
-        let new_count = self.task_ui.filtered_task_indices.len();
-        if new_count > 0 && self.task_ui.task_panel_index >= new_count {
-            self.task_ui.task_panel_index = new_count - 1;
-        }
-        self.refresh_task_view();
-    }
-
     /// Handle keys while the global-search strip is focused. Typed characters
     /// edit the query (so plain `j`/`k` insert, like the other search inputs);
     /// `Up`/`Down` and `Ctrl+P`/`Ctrl+N` move the selection; `Enter` activates
